+++ src/main/java/com/squeezecorp/pokermgm/dto/session/CreateSessionRequestDTO.java
@@ -1,4 +1,4 @@
-package com.squeezecorp.pokermgm.dto;
+package com.squeezecorp.pokermgm.dto.session;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,8 @@ import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
-public class UpdateSessionRequestDTO {
+public class CreateSessionRequestDTO {
+
     private String local;
     private LocalDateTime date_time;
     private Integer number_session;
