+++ src/main/java/com/squeezecorp/pokermgm/dto/session/UpdateSessionRequestDTO.java
@@ -1,4 +1,4 @@
-package com.squeezecorp.pokermgm.dto;
+package com.squeezecorp.pokermgm.dto.session;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,8 +7,7 @@ import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
-public class CreateSessionRequestDTO {
-
+public class UpdateSessionRequestDTO {
     private String local;
     private LocalDateTime date_time;
     private Integer number_session;
