+++ src/main/java/com/squeezecorp/pokermgm/service/session/SessionService.java
@@ -1,7 +1,7 @@
 package com.squeezecorp.pokermgm.service.session;
 
-import com.squeezecorp.pokermgm.dto.CreateSessionRequestDTO;
-import com.squeezecorp.pokermgm.dto.UpdateSessionRequestDTO;
+import com.squeezecorp.pokermgm.dto.session.CreateSessionRequestDTO;
+import com.squeezecorp.pokermgm.dto.session.UpdateSessionRequestDTO;
 import com.squeezecorp.pokermgm.model.Session;
 import com.squeezecorp.pokermgm.repository.SessionRepository;
 import jakarta.transaction.Transactional;
